move some includes and start work on a configurable session limit
There is no reason we should not centralize included files when
possible, so I moved the db and memcached session includes to the
framework.php file with all the other framework related includes.

I started to abstract the session lifetime value used for cookies,
but it's not functional yet and just part of an ongoing effort to
support a "stay logged in" option.